# Implement asynchronous `fill` method using `dpctl` kernels