--- a/dpnp/dpnp_algo/dpnp_fill.py
+++ b/dpnp/dpnp_algo/dpnp_fill.py
@@ -28,6 +28,7 @@ from numbers import Number
 
 import dpctl.tensor as dpt
 import dpctl.utils as dpu
+import numpy as np
 from dpctl.tensor._ctors import _cast_fill_val
 from dpctl.tensor._tensor_impl import (
     _copy_usm_ndarray_into_usm_ndarray,
@@ -60,9 +61,9 @@ def dpnp_fill(arr, val):
         )
         _manager.add_event_pair(h_ev, c_ev)
         return
-    elif not isinstance(val, Number):
+    elif not isinstance(val, Number) and not isinstance(val, np.bool_):
         raise TypeError(
-            f"Array cannot be filled with `val` of type {type(val)}"
+            f"array cannot be filled with `val` of type {type(val)}"
         )
 
     dt = arr.dtype