… tests/third_party/cupy/core_tests/test_ndarray_copy_and_view.py