@ndgrigorian 2024-09-17
@antonwolfy @vtavana @vlad-perevezentsev 
I've added a commit to skip `test_fill_with_numpy_scalar_ndarray` from the CuPy tests.

`dpnp.fill_diagonal` does not permit NumPy arrays, so it made sense to do the same here. If it would be preferred to keep this feature, it can be implemented, but I would argue that for consistency, the two should behave the same.